+++ src/language/csharpLanguage.ts
@@ -52,7 +52,7 @@ export const csharpLanguage: Language = {
           (attribute_argument_list
             (attribute_argument
               [
-                (verbatim_string_literal) 
+                (verbatim_string_literal)
                 (string_literal)
               ] @expression
             )